Because of stray console.log management. Needs better debugging workflow, I tell myself, that cleans up after itself. If only technology.
Note to commit readers - I’m thinking about writing a promise interface for jQuery promises. It may require rethinking and removing some of the callback based functionality as far as the API is concerned, but I’m working on it because I like promises much more than callbacks. It will probably be a separate file that you include that has a dependency on jQuery. The base file does not have external dependencies and will stay that way.

I love Anima, I really do - I use it in all web projects that I author. Author, take that, authors.